chore: compiler.webpack.moduleFilenameHelpers
